viq-transcriber/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
torch
soundfile
